fix conflicting types

+++ src/lib/carp_ht.h
@@ -16,7 +16,7 @@ typedef struct carp_ht_s {
 } carp_ht;
 
 void carp_ht_init (carp_ht *);
-int carp_ht_del (carp_ht *, char *);
+short int carp_ht_del (carp_ht *, char *);
 carp_ht *carp_ht_tail (carp_ht *);
 carp_ht *carp_ht_set (carp_ht *, char *, long long);
 carp_ht *carp_ht_get (carp_ht *, char *);
